feat: use the gates for the UserController system.

+++ app/Http/Controllers/Admin/UserController.php
@@ -4,6 +4,9 @@ namespace App\Http\Controllers\Admin;
 
 use Illuminate\Http\Request;
 use App\User;
+use App\Role;
+use Gate;
+use Illuminate\Http\Response;
 use Illuminate\Routing\Controller;
 
 class UserController extends Controller
@@ -11,7 +14,7 @@ class UserController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function index()
     {
@@ -20,81 +23,60 @@ class UserController extends Controller
         return view('admin.users.index')->with('users', $users);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        $request->validate([
-            'name'     => ['required'],
-            'email'    => ['required', 'email', 'unique:users'],
-            'password' => ['required', 'min:8'],
-        ]);
-
-        User::create([
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => bcrypt( $request->password),
-        ]);
-
-        return back();
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param User $user
+     * @return Response
      */
-    public function edit($id)
+    public function edit(User $user)
     {
-        //
+        if(Gate::denies('edit-users'))
+        {
+            return redirect(route('admin.users.index'));
+        }
+
+        $roles  = Role::all();
+
+        return view('admin.users.edit')->with([
+            'user' => $user,
+            'roles'=> $roles
+        ]);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @param User $user
+     * @return void
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, User $user)
     {
-        //
+        $user->roles()->sync($request->roles);
+
+        $user->name = $request->name;
+        $user->email = $request->email;
+        $user->save();
+
+        return redirect()->route('admin.users.index');
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param User $user
+     * @return Response
+     * @throws \Exception
      */
     public function destroy(User $user)
     {
+        if(Gate::denies('edit-users'))
+        {
+            return redirect(route('admin.users.index'));
+        }
+
         $user->delete();
 
         return back();
